Criar módulos individuais para cada endpoint
